refactoring (interfaces etc)
