+++ app/components/alumni/Alumni.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState, useMemo } from "react";
 import PlayerTable from "./PlayerTable";
-import { SelectedTeam } from "./TeamSearchBar";
+import { SelectedTeam } from "@/app/types/team"; 
 import { useFetchPlayers } from "./hooks/useFetchPlayers";
 import { RxMagnifyingGlass } from "react-icons/rx";
