+++ app/types/team.ts
@@ -21,4 +21,21 @@ export interface Team {
     flagUrl: string | null; // Added to store the flag URL for the player's nationality
   }
   
+  export interface TeamItem {
+    id: number;
+    name: string;
+    league: string;
+  }
+  
+  export interface SelectedTeam {
+    id: number;
+    name: string;
+  }
+  
+  export interface TeamSearchBarProps {
+    onSelect: (teamObj: SelectedTeam) => void; // optional single-select usage
+    onError: (err: string) => void;
+    selectedTeams: SelectedTeam[];
+    onCheckedTeamsChange: (teams: SelectedTeam[]) => void;
+  }
